requirements.txt
alembic==1.8.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coloredlogs==15.0.1
Flask==2.1.2
greenlet==1.1.2
humanfriendly==10.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pydantic==1.9.1
pyhumps==3.7.2
python-dotenv==0.20.0
requests==2.28.1
SQLAlchemy==1.4.39
typing_extensions==4.3.0
urllib3==1.26.9
Werkzeug==2.1.2
